+++ sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/core/execute/engine/ShardingExecuteGroup.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.execute;
+package org.apache.shardingsphere.core.execute.engine;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
