+++ sharding-core/sharding-core-execute/src/test/java/org/apache/shardingsphere/core/execute/sql/prepare/SQLExecutePrepareTemplateTest.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.core.execute.sql.prepare;
 
 import org.apache.shardingsphere.core.constant.ConnectionMode;
-import org.apache.shardingsphere.core.execute.ShardingExecuteGroup;
+import org.apache.shardingsphere.core.execute.engine.ShardingExecuteGroup;
 import org.apache.shardingsphere.core.execute.sql.StatementExecuteUnit;
 import org.apache.shardingsphere.core.route.RouteUnit;
 import org.apache.shardingsphere.core.route.SQLUnit;
